API/requirements.txt
google-generativeai
fastapi
python-multipart
uvicorn